- Changed scalers to be written as TScalerData to a ScalerTree. - Added --ignore_scaler option to suppress writing of ScalerTree (in the first 500MB subrun of run 4970 with 1ms dwell time, this reduces the fragment file size from 319MB to 226MB). - Added -Wall -Wextra -pedantic -Wno-unused-parameter  -Wno-unused-function as compiler options into the makefile and corrected the resulting warnings in the code.   - this was mainly unsigned-signed comparisons, unused parameters, and changing our color codes from \e to \033

+++ include/TDescantData.h
@@ -54,7 +54,7 @@ class TDescantData : public TGRSIDetectorData {
 		inline void SetDet(TFragment *frag,TChannel *channel,MNEMONIC *mnemonic)	{
 		   if(!frag || !channel || !mnemonic) return;
          
-         for(int x=0;x<frag->Charge.size();x++) {
+         for(size_t x=0;x<frag->Charge.size();x++) {
      		   SetDetNumber(mnemonic->arrayposition);
 	   	   SetDetAddress(frag->ChannelAddress);
             //SetDetEnergy(channel->CalibrateENG(frag->Charge.at(x)));
